brandonjplambert/scripts/requirements.txt
aiohttp>=3.8.0
pyyaml>=6.0
pandas>=1.5.0
openpyxl>=3.0.0
numpy>=1.21.0